chore(RT-82): rework modelBottomSheet widget
