+++ rentool/lib/features/map/view/map_delivery_screen.dart
@@ -63,12 +63,30 @@ class _MapDeliveryScreenState extends State<MapDeliveryScreen> {
               onMapCreated: (controller) {
                 _mapControllerCompleter.complete(controller);
               },
-              onMapLongTap: (point) async {
-                BlocProvider.of<MapBloc>(context).add(MapUpdatePointEvent(
-                    point: MapPoint(
-                        address: 'Неизвестный адресс',
-                        latitude: point.latitude,
-                        longitude: point.longitude)));
+              onMapTap: (point) async {
+                final mapBloc = BlocProvider.of<MapBloc>(context);
+                final completer = Completer();
+
+                mapBloc.add(MapUpdatePointEvent(
+                  point: MapPoint(
+                      address: 'Неизвестный адресс',
+                      latitude: point.latitude,
+                      longitude: point.longitude),
+                  completer: completer,
+                ));
+
+                await completer.future;
+
+                showModalBottomSheet(
+                  // ignore: use_build_context_synchronously
+                  context: context,
+                  builder: (context) => ModalBodyView(
+                    point: state.point,
+                    mainText: 'Укажите адрес доставки',
+                    buttonText: 'Сохранить адрес',
+                    receivingMethod: ReceivingMethods.delivery.value,
+                  ),
+                );
               },
               mapObjects: mapObjects,
             );
@@ -111,106 +129,7 @@ class _MapDeliveryScreenState extends State<MapDeliveryScreen> {
       icon: PlacemarkIcon.single(PlacemarkIconStyle(
           image: BitmapDescriptor.fromAssetImage(
               'assets/icons/Yandex_Maps_icon_min.png'))),
-      onTap: (_, __) => showModalBottomSheet(
-        context: context,
-        builder: (context) => _ModalBodyView(point: point),
-      ),
+      onTap: (_, __) => (),
     );
-  }
-}
-
-class _ModalBodyView extends StatefulWidget {
-  const _ModalBodyView({required this.point});
-
-  final MapPoint point;
-
-  @override
-  State<_ModalBodyView> createState() => _ModalBodyViewState();
-}
-
-class _ModalBodyViewState extends State<_ModalBodyView> {
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(vertical: 50),
-      child: SizedBox(
-        width: double.infinity,
-        child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 24),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            mainAxisAlignment: MainAxisAlignment.start,
-            children: [
-              const Row(
-                children: [
-                  Text(
-                    'Укажите адрес доставки',
-                    style: TextStyle(
-                      fontSize: 22,
-                      color: Colors.black,
-                      fontWeight: FontWeight.w700,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 16),
-              const Row(
-                children: [
-                  Text(
-                    'Адрес',
-                    style: TextStyle(
-                      fontSize: 12,
-                      color: Color(0xfffaaaaaa),
-                      fontWeight: FontWeight.w400,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 8),
-              Row(
-                children: [
-                  Text(
-                    widget.point.address,
-                    style: const TextStyle(
-                      fontSize: 12,
-                      color: Colors.black,
-                      fontWeight: FontWeight.w400,
-                    ),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 30),
-              ButtonPrimary(
-                onPressed: _onPressed,
-                text: 'Сохранить адрес',
-                style: const TextStyle(
-                  fontSize: 14,
-                  color: Colors.black,
-                  fontWeight: FontWeight.w500,
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  Future<void> _onPressed() async {
-    final rout = context.router;
-    final mapBloc = BlocProvider.of<MapBloc>(context);
-    final completer = Completer();
-
-    mapBloc.add(MapAddAddressEvent(
-      point: widget.point,
-      type: 1,
-      completer: completer,
-    ));
-    await completer.future;
-
-    mapBloc.add(MapLoadAddressEvent());
-
-    await rout.maybePop();
-    await rout.maybePop();
   }
 }
